requirements.txt
# AI-Driven Malware Sandboxing System - Requirements
# Core Data Science Libraries
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine Learning Libraries
scikit-learn>=1.0.0
imbalanced-learn>=0.8.0

# Feature Engineering & Analysis
scipy>=1.7.0
joblib>=1.0.0

# JSON Processing & API Handling
requests>=2.25.0
urllib3>=1.26.0

# Configuration & Logging
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Progress Bars & Utilities
tqdm>=4.62.0
colorama>=0.4.4

# Cuckoo Sandbox Integration
pymongo>=3.12.0  # For Cuckoo database if using MongoDB
psycopg2-binary>=2.9.0  # For PostgreSQL database

# Security & Hashing
hashlib-compat>=1.0.0
cryptography>=3.4.0

# Visualization
plotly>=5.0.0
dash>=2.0.0  # For web dashboard (optional)

# Testing & Development
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0

# Windows-specific dependencies
pywin32>=301; sys_platform == "win32"
wmi>=1.5.1; sys_platform == "win32"
